src/requirements.txt
langgraph
langchain-core
langchain-community
trustcall
watchdog
langchain
langchain-community
langchain-groq
langchain-huggingface
python-dotenv
pydantic
pydantic-settings
chromadb
sentence-transformers
watchdog
pypdf
docx2txt
httpx
pandas
IPython
psycopg_pool
langchain-postgres>=0.0.12
pgvector>=0.3.0
psycopg[binary]>=3.1.0
langchain_postgres
psycopg2-binary==2.9.9
sqlalchemy
langgraph[postgres]
langgraph-checkpoint-postgres